Move caret behind newly inserted links
Fixes #6711.

This allows the user to continue typing after inserting a link, without
having to fiddle with the caret position to get it out of the <a> (which
takes two presses of the right arrow key, but that's hard to guess).
This behavior matches Google Docs.

This affects links inserted through the `link` plugin's dialog, the
`quickbars` plugin's hovering UI, as well as the `paste` plugin's
SmartPaste feature. I've tested those three cases, and the new behavior
seems desirable in each case. It seems like, after inserting a URL onto
text, a user *usually* does not want to edit it further, but continue
typing.